BACKPORT: e1000e: Remove special case for 82573/82574 ASPM L1 disablement
commit 59aed95263bdd0e2b48eb9be5a94346d2d4abf90 upstream.

For the 82573, ASPM L1 gets disabled wholesale so this special-case code
is not required. For the 82574 the previous patch does the same as for
the 82573, disabling L1 on the adapter. Thus, this code is no longer
required and can be removed.

Signed-off-by: Chris Boot <bootc@bootc.net>
Tested-by: Jeff Pieper <jeffrey.e.pieper@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(cherry picked from commit e60a87bab7ce339c034b7d7dd365d687bbffd091)

BUG=chromium-os:35241
TEST=Checked connectivity on a board with an i217 Ethernet controller

Change-Id: Id20cb1076f5cd3ed349b33d654ea6b253c84da8d
Reviewed-on: https://gerrit.chromium.org/gerrit/36105
Commit-Ready: Olof Johansson <olofj@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>