CHROMIUM: kick lockup detector on resume
On the suspend/resume path the boot CPU does not go though the
offline->online transition. This breaks the NMI detector
post-resume since it depends on PMU MSR state that gets
lost when the system gets suspended.

As a workaround, force a CPU offline->online transition for the
lockup detector on the boot CPU during resume.

I still need to take this upstream but wanted to get this checked
into the Chrome OS kernel repo first since we are getting significant
number of system freezing reports from the field. We have no good way
of debugging these without the stack trace that the NMI watchdog
generates in such situations.

BUG=chrome-os-partner:9303
TEST=do a suspend resume cycle and then issue this command:
taskset 1 sh -c "echo nmilockup > /proc/breakme". The system should
reboot in about 10 seconds.

Change-Id: Ica9b206fa8b1b96475f136e31068a0a7621c758f
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21263
Reviewed-by: Olof Johansson <olofj@chromium.org>