CHROMIUM: Work around Intel GPU bug with some physical pages.
Some intel GPUs don't like using those physical pages, resulting in
garbage all over the place. We could fix it by adding an e820 region,
but that prevents the machine from booting. We could fix it in i915,
but since the allocation is done by the backing shmem it's not
possible. So we have to do the workaround in shmem.c.

As a side note, if we have that workaround in place, we can probably
reclaim 4 megabytes of memory (twice 2 megs) which are marked
reserved by the bios.

BUG=chromium:120949
TEST=by hand; set mem=450M and play with Chrome OS, corruption is gone.

Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>

Change-Id: I957e125fb280e0b0d6b05a83cc4068df2f05aa0a
Reviewed-on: https://gerrit.chromium.org/gerrit/22308
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>