CHROMIUM: drm/i915/intel_i2c: always wait for IDLE before clearing NAK
The GMBUS controller can report a NAK condition while a transaction is
still active. If the driver is fast enough, and the bus is slow enough,
the driver may clear the NAK condition while the controller is still
busy, resulting in a confused GMBUS controller.  This will leave the
controller in a bad state such that the next transaction may fail.

Also, return -ENXIO if a device NAKs a transaction.

Note: this patch also refactors gmbus_xfer to remove the "done" label.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:26118
TEST=1) enable DRM_DRIVER_DEBUG:
    echo -n 2 > /sys/module/drm/parameters/debug
  2) Generate a fast, endless stream of NAKs on vga port using:
    while true; do i2cdetect -y 2 0x50 0x50 ; done
  3) Notice GMBUS NAK messages in dmesg
     [drm:gmbus_xfer], GMBUS [i915 gmbus vga] NAK for addr: 0050 w(0)
  4) At the same time, move finger on gmbus-connected trackpad.
  => Reads from trackpad should not fail with "IDLE" or "Timeout" errors.

Change-Id: I6e7fa3191af0f98368029e2783d20f8e67724c75
Reviewed-on: https://gerrit.chromium.org/gerrit/18383
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>

Conflicts:

	drivers/gpu/drm/i915/intel_i2c.c