CHROMIUM: drm/i915/intel_i2c: use INDEX cycles for i2c read transactions
It is very common for an i2c device to require a small 1 or 2 byte write
followed by a read.  For example, when reading from an i2c EEPROM it is
common to write and address, offset or index followed by a reading some
values.

The i915 gmbus controller provides a special "INDEX" cycle for performing
such a small write followed by a read.  The INDEX can be either one or two
bytes long.  The advantage of using such a cycle is that the CPU has
slightly less work to do once the read with INDEX cycle is started.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:26118
TEST=a) EDID read from hdmi & panel (LVDS) (check /var/log/Xorg.0.log)
TEST=b) atmel_mxt_ts driver can detect touchpad on vga ddc, and process
     touch events.

Change-Id: Ic5817cb1de98d1d793c01814d7a953aa487c4c0c
Reviewed-on: https://gerrit.chromium.org/gerrit/17935
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>

Conflicts:

	drivers/gpu/drm/i915/intel_i2c.c