BACKPORT: ALSA: hda - Always resume the codec immediately
This is a fix for the problem in commit 785f857d1c, the pop noise
issue on some machines with ALC269.  The problem was the uninitialized
state after the resume due to the delayed resume of the codec chips.
In that commit, we tried to fix by forcibly putting the codec to D3 at
suspend.  But, this still also leaves the uninitialized state after
resume, and it _might_ be still problematic with some BIOS.  Since the
commit turned out to regress another issues, we reverted it in the
end.

Now, in this fix, try to fix by turning on the codec immediately at
the resume path.  We need to take care of the power-saving in this
case.  When the device is woken up at the power-saved state, it should
go power-saving again after the resume.

BUG=chrome-os-partner:10399
TEST=suspend and resume hda codec.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25321
Reviewed-by: Olof Johansson <olofj@chromium.org>
(cherry picked from commit 7f30830b7b82e5225c38a48b387e44f3defe40e2)

Change-Id: I7457a47f62a3de1a3fd6274afa508bb35e4c71d9
Reviewed-on: https://gerrit.chromium.org/gerrit/25628
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>