CHROMIUMOS: mac80211: Use different probe counts on resume
We recently increased the probe retry count for beacon monitoring
in order to make challenging channel conditions less volatile.  In
doing so, we made the system slower to respond to a lost AP when the
system resumes from suspend.  Mitigate this by using a different
try-count for the resume-from-suspend case.

BUG=chromium-os:28576
TEST=Manual: Suspend, unplug AP, resume.  Observe disconnection in < 1sec

Change-Id: I206d4b2838da2ce7da00739412e5fcda63a27a2c
Signed-off-by: Paul Stewart <pstew@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/19234
Reviewed-by: Gary Morain <gmorain@chromium.org>

Conflicts:

	net/mac80211/mlme.c