ALSA: hda/realtek - set DMIC coefficients
The DMIC coefficients got lost in the merge from 3.0 to 3.2.  Without
them the right channel is exactly the inverse of the left.  GTalk sums
the two channels to get a mono stream, leading to zeros.  Use the new
chained fixup method to apply the needed coefficients.

BUG=chromium-os:29356
TEST="arecord -Dhw:1 -f cd /tmp/asdf.wav" copy file to host and use
audacity (or similar) to check that signals wouldn't add to zero.
GTalk call and ensure other party can hear you.

Change-Id: I051dce52866c7d6119e12b1489b79fa04fd543fa
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/22440
Reviewed-by: Todd Broch <tbroch@chromium.org>