video: msm: Clear previous histogram data on next startup
Clear the pointer to the userspace histogram struct when histogram is
started. Previously, when the histogram was stopped while a histogram read
was in progress, the pointer was not cleared, causing the next histogram
read to think that someone was already using the histogram.

CRs-Fixed: 361640
Signed-off-by: Carl Vanderlip <carlv@codeaurora.org>
(cherry picked from commit 19e4b71f617e307cacf0be077e08bec5c37df290)

Change-Id: Iaac2ff19bd3559ef602cd1a21a3f44e9a64e9f7d
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>