CHROMIUMOS: mac80211: Increase retry attempts for nullfunc probes
When a nullfunc probe fails for lack of an ACK, it is retried
almost immediately.  The current setting of 10 retries completes
in about 20 ms.  If an AP is off-channel scanning, this is not
enough time for the AP to return to the operating channel and
ACK the nullfunc.  As a result, the client disconnects when the
AP is still there and operating.

BUG=chrome-os-partner:8694
TEST=manual.  Let run overnight to verify that 100 is a sufficient
number of retries.  50 is too low.

Change-Id: I0e8a7032fe50a2aa2628edcdd52330dac06e6eb9
Signed-off-by: Gary Morain <gmorain@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/24652
Reviewed-by: Paul Stewart <pstew@chromium.org>