UPSTREAM: ath9k: Handle fatal interrupts properly
When a fatal interrupt is received or it is detected that the baseband
has hung, the chip has to be reset immediately.  Otherwise, we end up
processing spurious interrupts. Ensure that we bail out properly in
the ISR when the reset work hasn't completed yet.

Signed-off-by: Sujith Manoharan <c_manoha@qca.qualcomm.com>
Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chromium-os:31625
TEST=Recompile, install, join GoogleGuest

Change-Id: I50470083dc66b8a19241eb928b96fef46366d492
Reviewed-on: https://gerrit.chromium.org/gerrit/24796
Reviewed-by: Gary Morain <gmorain@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>