CHROMIUMOS: config: Enable verbose logging in mac80211
Enable CONFIG_MAC80211_VERBOSE_DEBUG, which causes the wireless
mac80211 module to emit more debug info into the syslog.  This
is necessary so that connectivity issues can be correctly
diagnosed from the data provided by the feedback tool.

BUG=chromium-os:31145
TEST=manual.  Build a kernel, load it onto Alex, and create a
"connectivity event" by unplugging the AP to which the machine
is connected.  Use the "Report An Issue" tool and look at the
system information, particularly the syslog, to see kernel events
that are under the CONFIG_MAC80211_VERBOSE_DEBUG compile flag.

Change-Id: I32e0985df0d2f584944b77e4e7949d9de1686d8a
Signed-off-by: Gary Morain <gmorain@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/23162
Reviewed-by: Paul Stewart <pstew@chromium.org>