CHROMIUM: config: base.config: enable pl2303 driver
Enable pl2303 driver for all platforms so that we could emulate
USB-to-RS232 converters when needed. Note that only the factory
test images will be installed with this driver. Two CLs for that
is following.

BUG=chrome-os-partner:6616,chrome-os-partner:6617
TEST=emerge chromeos-kernel -> got pl2303.ko
tested the ko file on link with converter devices -> works

Change-Id: I9c8e6f2f6c65341e0890f3102385ec6552fc3fdd
Signed-off-by: Tai-Hsu Lin <sheckylin@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/23848
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/24294
Commit-Ready: Rong Chang <rongchang@chromium.org>
Reviewed-by: Rong Chang <rongchang@chromium.org>
Tested-by: Rong Chang <rongchang@chromium.org>