CHROMIUM: platform/x86/chromeos_laptop - mxt224s palm settings
Make the following changes to the mxt224s touchpad settings:
T9.08 = 4 (to reduce the confetti effect during palm contact)
T9.10 = 5 (to enable MRGHYST to take effect)
T9.15 = 5
T9.16 = 15

Summary :
These changes target more stable palm identification on the mxt224SL
touchpad. Palms should be reported as a single large pressure contact,
and not as "confetti" or "fireworks." These settings attempt to balance
good palm identification with finger seperation for two finger gestures.
Increase Touch Detect Integration field to 4. Enable Merge Timeout with
a timeout of 800ms.
Set merge threshold to 5 and merge hysteresis to 15.

BUG=chrome-os-partner:9737
TEST=use DISPLAY=:0.0 mtplot
Bring down a palm on the top of the touchpad in a natural manner as if
you are resting in a ready position for the keyboard. Verify that
the driver reports a single large touchpoint for the palm, and that it
doesn't "confetti."
Verify that fat close fingers perform well in 2+ finger scroll gestures.

Change-Id: Ide52aa0c3ad361380fdb85bddea25ee55b29e9b1
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/24547
Reviewed-by: Olof Johansson <olofj@chromium.org>