msm_fb: HDMI: Update HPD logic to address HDMI PLL related issues
In the current implementation, the regulators pertaining to the
HDMI core are turned off/on as part of the configuration of the
HDP circuitry. As a result, the regulator that powers the HDMI
PLL is turned off before the HDMI clocks on the PLL are disabled.
This might lead to the PLL not getting locked when it is re-enabled.

This change turns on and off the regulators and clocks in proper
sequence.

CRs-Fixed: 360135
CRs-Fixed: 358598
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>
Signed-off-by: Ajay Singh Parmar <aparmar@codeaurora.org>
(cherry picked from commit 7d11c271b2ca399783975d3b6ef3199cd937a2e0)

Change-Id: I7d65d0fc71ec2178b66bd22bea60168cbd1aeff5
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>
(cherry picked from commit 55068d2be4d29a82bbf20da8ee67e07922fd95ae)

Signed-off-by: Dhivya Subramanian <dthiru@codeaurora.org>