CHROMIUM: config: build in atmel_mxt_ts driver
Build in the atmel_mxt_ts driver to speed up initialization
of atmel_mxt touch devices. Loading as module results in the
input device not responsive for some seconds as the rest of boot
happens quite quickly.

Bulding it in will improve usability, and also allow us to
perform firmware and config updates early on at boot time.

BUG=chrome-os-partner:9720
TEST=Boot the system. Check for atmel_mxt_ts device initialization.
Check that it occurs around 3s.

Change-Id: I7dfd9b76aed9f4fbdf60c6b7b9bc86b0f44934e9
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/24830
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>