CHROMIUM: platform/x86/chromeos_laptop - probe for atmel_mxt_tp bl
Use chromeos_laptop_add_probed_i2c_device to handle atmel_mxt_tp
devices that start in bootloader mode, a condition that may
happen when there is corrupted firmware, but a functional bootloader,
for example.

This will allow the device to be initialized with the bootloader address
and a recovery to occur.

BUG=chrome-os-partner:10318
TEST=Perform a firmware update on the atmel_mxt_tp device, and before
it completes, reboot the system to leave the touchpad in an incomplete
operational firmware state.
On the next boot, the atmel driver should be loaded on 0x25,
so you should find an input device at /sys/bus/i2c/devices/1-0025
Perform another uninterrupted update with chromeos-touch-firmwareupdater.sh
and it should recover.

Change-Id: I11e7e8d6d61b7d5b731f388ef655c8d28ec122e6
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/24832
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>