CHROMIUM: drivers: power: Add TPS65090 battery charger
This driver reads AC status from the tps65090 directly.

Note: At present the TPS65090 has a few problems, one of which is that
the status registers do not work. As a result we are reading registers
that we should not. Definitions for these registers have been put into
tps65090-private.h for now.

BUG=chrome-os-partner:10617,chrome-os-partner:11739,chrome-os-partner:12244
TEST=Build successfully

Change-Id: I909fb5919cb782c49a72c488b3b36b245b029112
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31042
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>