CHROMIUM: power: tps65090: Add charger regulator and device tree support
Add a simple fdt binding to this driver. Also use a static struct to
provide the required power supply detail instead of assigning them at
run-time.

If the fdt node is not present or is disabled, we do not register the
driver. If the charger sub-node is not present, we do not register the
regulator, and operate entirely in passive mode (as the driver always
did before this change).

BUG=chrome-os-partner:10617
TEST=build and boot to kernel prompt on snow

Change-Id: I93ec83b44f53f497d1adf745ddb45e2d438402e4
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40166
Reviewed-by: Benson Leung <bleung@chromium.org>