CHROMIUM: charger-manager: Allow battery to provide temperature
In some systems the we cannot obtain the cell temperature using the charger
or an ADC; we must ask the battery (fuel cell) since it has thermistors
which are closer to the battery and thus give a more accurate reading.

Add this feature to charger-manager.

BUG=chrome-os-partner:10617
TEST=build and boot to kernel prompt on snow
See that we obtain the correct temperature now (from the battery)

Change-Id: I2f2844632b7ef7b1179878cc769bcbe751c4c708
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40169
Reviewed-by: Sameer Nanda <snanda@chromium.org>