USB: msm72k_udc: Add counter to reduce delay required for qTD status update
HSUSB PHY is taking time to update the dTD status(INFO_ACTIVE) of the
queued request. So this is checked after a 10us delay, as a data stall
issue is seen without this delay.

Time to get dTD status updated is not deterministic and adding 10us delay
is reducing FTP download throughput. Hence check the status of dTD every
1us for 10 times instead of waiting for 10us.

CRs-Fixed: 371338
Change-Id: Ied0a0792b38a11835df058a187e8974a7cd6671b
Signed-off-by: Rajkumar Raghupathy <raghup@codeaurora.org>