CHROMIUM: config: Increase hda power save timeout.
Increase the HDA power saving timeout to 15 seconds.  Pepper flash polls
the sound interface every 5 seconds.  Having the power-down timeout the
same as this interval causes the polls to have to wait for the power
down to complete.  While Chrome waits for this it can't send messages to
the GPU process, causing a lot of UI lag.

This change is temporary, bug 32055 filed to track undoing this.

BUG=chromium-ow:32037
TEST=Open gmail tab and a scratchpad and make sure that the rendering of
keys isn't interrupted for long periods of time.

Change-Id: I9ac533c5a4aecf020c2f794d1fc8d67ac6e74f75
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25904
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Olof Johansson <olofj@chromium.org>