mako: Disable pull-down in requests for S7 and L27 regulators
msm: board-8064: Disable pull-down in requests for S7 and L27 regulators
commit	fd487fdd37fab8fb6bdda6fce9a58cee2e041384

Change the configuration of PM8921 regulators SMPS 7 and LDO 27
so that requests sent to the RPM for these regulators have pd=0
(i.e. pull-down disabled).  This modification reduces leakage
current that is seen during XO shutdown when S7 and L27 are
disabled but S3 is kept on.  The outputs of S3 and L27 are wired
together so the S3 output voltage is continuously dropped across
the L27 pull-down resistor if it is enabled.  S7 is the parent
supply of L27 so the same problem is seen if its pull-down
resistor is enabled.

Change-Id: I361bb3017221d7a5b920549af03c5fa437a221aa
CRs-Fixed: 369947