msm: gdsc: Delay 1us after enabling rails while clocks are re-enabled
The GDSC state-machine will temporarily disable clocks to a power
domain while the core reset sequence is performed. After power is
enabled, the clocks will take 4 cycles to re-enable.

This scenario only occurs if gdsc_enable() is called while clocks are
enabled. While this is not the recommended order for device drivers
to perform these steps, the 1us delay is added here just in case
the recommendation is not followed.

Change-Id: Ieff39755561b6f82731747cf0984d8d401001060
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>