CHROMIUM: config: add a fully populated exynos5 config
This adds a fully populated (not yet split-configged) exynos5
config from the previously private tree. I'm importing this
as a separate commit to make it easier to review actual contents.

Note that this is imported as a chromeos config, not chromiumos config.
The chromiumos counterpart will be removed once the public builder has been
moved over to the other config.

BUG=chrome-os-partner:10937
TEST=none

Change-Id: I35c8b00d5242400471ce0114598f11ed614b475c
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26287
Reviewed-by: Pawel Osciak <posciak@google.com>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>