CHROMIUM: config: renormalize splitconfig
Normalizing splitconfig and turning off USB_GADGET, MTD and switching some of the
v4l options to modules.

This also switches CONFIG_MEDIA_CONTROLLER to =m on exynos to keep it common with
other platforms.

Finally, it also copies over the chromeos-exynos5 config to
chromiumos-exynos5. The old tegra placeholder is no longer needed,
we can do the real thing.

BUG=chrome-os-partner:10937
TEST=none

Change-Id: Ic4bf0e27cb8291ee5b88cd27af7851aaf98855ce
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26288
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>