CHERRY-PICK: exynos: i2c: Add bus arbitration implementation
The arbitrator is a general purpose function which uses two GPIOs to
communicate with another device to claim/release a bus. We use it to
arbitrate an i2c port between the AP and the EC.

BUG=chrome-os-partner:10955
TEST=boot on snow; see that i2c bus can still be accessed, and that
EC has to wait sometimes. Keyboard operates while the EC is talking to
the battery and PMU.

Torture test:

On AP:
while true; do i2cdump  -f -y 4 0x48; done >/dev/null

On EC:
pmu 10000

This runs the bus flat out in with both masters for about 9 minutes.
very roughly 150k i2c transactions on each side.

Observed that we got only 3 timeouts on the EC in that time, none
on the AP, and that the EC recovered from timeout events correctly.

Change-Id: Ib78d3a418c2c97f78e7d97ce94b446749747c872
Original-Change-Id: I9dc9406f68b40bfbdec001514775731fa195a714
Original-Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27122
Commit-Ready: Katie Roberts-Hoffman <katierh@chromium.org>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Tested-by: Katie Roberts-Hoffman <katierh@chromium.org>