msm: clock-local2: Modify halt check to account for NOC FSM handshake
When enabling certain branch clocks, there may be a handshake
performed with an FSM that indicates that the slave off of a
NOC is ready for traffic from that NOC. Modify the branch
clock halt checking to account for this additional handshake.
Note that only a few clocks have this requirement, and hence
the previously used halt check must still be preserved.

Change-Id: I0cdb454e5a89444e3ed9edadb88a7bc5a88c7a99
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>