CHROMIUM: gsmi: enable CONFIG_GOOGLE_GSMI for x86 devices
This enables kernel shutdown messages in the event log on supported
platforms.  The driver will not get loaded on unsupported platforms
as it is guarded by a SMBIOS board or BIOS vendor whitelist.

The driver no longer depends on EFI_VARS so it does not cause an
unrelated panic on systems with 64bit EFI and 32bit kernel.

Configs were re-generated with kernelconfig to ensure that no
additional dependencies are added.

BUG=chrome-os-partner:9195
TEST=manual test on link/x86_64 and alex/i686

1) Boot kernel on 64bit link and 32bit alex devices
2) Look for GSMI messages on supported link device:
localhost kernel: [    3.283713] gsmi version 1.0 loaded
3) Add an event and see that it shows up in the log:
> xxd -g 1 test-event-kernel_panic
0000000: 01 00 00 00 ad de 02 00 00 00
> cat test-event-kernel_panic > /sys/firmware/gsmi/append_to_eventlog
> mosys eventlog list | tail -1
55 | 2012-06-29 14:33:07 | Kernel Event | Panic

Change-Id: Ia66490cf46a405e9aa2cc81dadc421cd6d47c778
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26490
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>