CHERRY-PICK: CHROMIUM: ARM: exynos: add flip traces
While we generally plan to expand this set of traces to cover more of
the driver, these two for flip request/complete are 2 of the 3 traces
currently reported by the i915 driver being used as a point of
comparison (the other is for the pwrite_ioctl which isn't supported
here).

TEST=Booted the kernel on device and was able to observe believable
trace data from systrace (once it was set to whitelist these
tracepoints).
BUG=chrome-os-partner:11092

Change-Id: Ie1dd6d6f6f82753f6b14a1cc5d8674842b274c74
Original-Change-Id: I70f2f77581f83ab0af83bdb891ffba51cd51e8e0
Original-Signed-off-by: Jeff Disher <disher@google.com>
Original-Reviewed-on: https://gerrit.chromium.org/gerrit/26721
Original-Reviewed-by: Sam Leffler <sleffler@chromium.org>
Original-Commit-Ready: Jeff Disher <disher@chromium.org>
Original-Tested-by: Jeff Disher <disher@chromium.org>
Original-Reviewed-by: Anush Elangovan <anush@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/27119
Commit-Ready: Katie Roberts-Hoffman <katierh@chromium.org>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Tested-by: Katie Roberts-Hoffman <katierh@chromium.org>