CHROMIUM: ARM: exynos: add flip traces
While we generally plan to expand this set of traces to cover more of
the driver, these two for flip request/complete are 2 of the 3 traces
currently reported by the i915 driver being used as a point of
comparison (the other is for the pwrite_ioctl which isn't supported
here).

TEST=Booted the kernel on device and was able to observe believable
trace data from systrace (once it was set to whitelist these
tracepoints).
BUG=chrome-os-partner:11092

Change-Id: I70f2f77581f83ab0af83bdb891ffba51cd51e8e0
Signed-off-by: Jeff Disher <disher@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/26721
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Commit-Ready: Jeff Disher <disher@chromium.org>
Tested-by: Jeff Disher <disher@chromium.org>
Reviewed-by: Anush Elangovan <anush@google.com>