UPSTREAM: ARM: 7445/1: mm: update CONTEXTIDR register to contain PID of current process
This patch introduces a new Kconfig option which, when enabled, causes
the kernel to write the PID of the current task into the PROCID field
of the CONTEXTIDR on context switch. This is useful when analysing
hardware trace, since writes to this register can be configured to emit
an event into the trace stream.

The thread notifier for writing the PID is deliberately kept separate
from the ASID-writing code so that we can support newer processors using
LPAE, where the ASID is stored in TTBR0. As such, the switch_mm code is
updated to perform a read-modify-write sequence to ensure that we don't
clobber the PID on CPUs using the classic 2-level page tables.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
(cherry picked from commit 575320d625d5b5eb115575a1f5e17af456e69577)

BUG=chromium-os:34007
TEST=general system testing

Change-Id: I91ad3d3972d36ef180b43a84f2abaf252b17080c
Original-Change-Id: I4a4ba8b8911efac06914ed0c2974d16b502d78f2
Signed-off-by: Olof Johansson <olofj@chromium.org>
Original-Reviewed-on: https://gerrit.chromium.org/gerrit/31715
Original-Reviewed-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31857
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>