CHROMIUM: cpufreq_interactive: merge up/down work
Down-clocking was being done in a workqueue but there is already a
realtime task for handling up-clocking: do both in the task and remove
the workqueue.  This eliminates conflicts and the mtx that guarded
frequency/clock changes.  As a side-effect input handler registration
is now done on the system workqueue.

While here also move the updown state mtx so that pcpu data
are correctly synchronized between cpufreq_interactive_boost and
cpufreq_interactive_timer (handling cpufreq_interactive_updown_task is
more difficult and may not matter).

Signed-off-by: Sam Leffler <sleffler@chromium.org>

BUG=none
TEST=run workloads and examine event traces

Change-Id: I84937df032287c925499d88bbdd2931f213ac1b9
Reviewed-on: https://gerrit.chromium.org/gerrit/27096
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Ready: Sam Leffler <sleffler@chromium.org>