msm: mdm: Add support for remote MDM image upgrade
mdm-driver is informed that an image upgrade of the remote modem is
required, at which point it stops monitoring the error gpios and takes
the appropriate actions. Support for the "USB Switch" gpio is added.
This is needed to write a image to an empty modem flash.

Change-Id: If390a82917200164f2c075142c0b511c9787659b
Signed-off-by: Ameya Thakur <ameyat@codeaurora.org>
Signed-off-by: Joel King <joelking@codeaurora.org>
(cherry picked from commit c4c41100d2eb23ae1fad606b7c6110d2b1b726c2)