tty: n_smux: Fix packet close synchronization issue
If a channel is logical channel open or close ack is pending in the
transmit queue when a logical channel is closed, then they are currently
purged and the remote and local channel states are mismatched and the
port cannot be re-opened.  Note that this issue only occurs during rapid
open/close/open sequences that are typically only seen in remote
loopback stress testing.

Add code to not purge ACK commands from a logical channel unless the
logical channel purge is due to a subsystem restart (at which point the
remote state is known to be power-on default).

(cherry picked from commit 0e91408cb6343d5ac18206ff3d6639ccec5464fd)

Change-Id: I87c46f278f4789235e31f586c9c47bd12b65f683
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>
Signed-off-by: Jessica Gonzalez <jgaona@codeaurora.org>