CHROMIUM: Input: atmel_mxt_ts - Remove input_free_device in create
In mxt_input_dev_create, if data->input_dev exists, we unregister
and then free the input device. This may be incorrect, as the unregister
should take care of freeing the input device when there are no longer any
references to the device.

BUG=chromium-os:32512
TEST=builds clean.
Use update_fw, and verify the input device is functional
after the firmware update.

Change-Id: Ic112803243c0e1904f7c8069693f9d0d17511482
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27250
Reviewed-by: Yufeng Shen <miletus@chromium.org>