CHERRY-PICK: UPSTREAM: ARM: 7467/1: mutex: use generic xchg-based implementation for ARMv6+
The open-coded mutex implementation for ARMv6+ cores suffers from a
severe lack of barriers, so in the uncontended case we don't actually
protect any accesses performed during the critical section.

Furthermore, the code is largely a duplication of the ARMv6+ atomic_dec
code but optimised to remove a branch instruction, as the mutex fastpath
was previously inlined. Now that this is executed out-of-line, we can
reuse the atomic access code for the locking (in fact, we use the xchg
code as this produces shorter critical sections).

This patch uses the generic xchg based implementation for mutexes on
ARMv6+, which introduces barriers to the lock/unlock operations and also
has the benefit of removing a fair amount of inline assembly code.

BUG=chrome-os-partner:13943
TEST=Booted

Change-Id: Id2523901dd2179c4092a3210c3602c6a074e571f
Original-Change-Id: Idf504e4778be7cdc50c5f1a121fe24c53a9746b7
Cc: <stable@vger.kernel.org>
Acked-by: Arnd Bergmann <arnd@arndb.de>
Acked-by: Nicolas Pitre <nico@linaro.org>
Reported-by: Shan Kang <kangshan0910@gmail.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from commit a76d7bd96d65fa5119adba97e1b58d95f2e78829)
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Pawel Osciak <posciak@google.com>
(cherry picked from commit a47581561741e349250fc29e30e44daf972f9897)
Reviewed-on: https://gerrit.chromium.org/gerrit/33370