msm: clock: Remove some clk_prepare/unprepare() warnings
Remove the warnings that the clk_enable() and clk_disable() APIs
print when drivers attempt to enable or disable clocks that are
not prepared. These warnings are currently harmless, but greatly
contribute to noise in the kernel logs.

This is safe because use of clk_prepare() and clk_unprepare() is
currently optional. They are placeholder APIs that do not
currently perform any operations. Their use will eventually be
required in future kernels, but never on this kernel branch.

Change-Id: Ib0a47a7402c7013605b89c840b0fcf478fd7a0af
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>