CHROMIUM: gobi: use IMEI when MEID is not available
The previous CL (https://gerrit.chromium.org/gerrit/11700) changes the
Gobi kernel driver to return an error if MEID is not available. That
prevents the driver from supporting modems without a MEID. This CL
changes the driver to obtain both MEID and IMEI from a modem and use
IMEI when MEID is not available.

BUG=chromium-os:32690
TEST=Tested the following:
1. emerge-{x86,amd64,arm}-generic chromeos-kernel
2. Verify that the driver handles modems with and without a MEID.

Change-Id: I2624f22022e406ebfd3d6dea767fa7a3b3944dfc
Signed-off-by: Ben Chan <benchan@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27784
Reviewed-by: Paul Stewart <pstew@chromium.org>