msm: mdm: change the gpio pull value of mdm2ap_status after bootup
The mdm2ap_status gpio has been known to fluctuate in
its value during mdm bootup if set to PULL_NONE, so set
it to PULL_DOWN when the gpiomux is initialliy installed.
However, it needs to be reset to PULL_NONE after the mdm
is booted up in order to reduce leakage current.

CRs-Fixed: 368433
CRs-Fixed: 379343
Change-Id: Id42aea731140e64800576ead73e800b18fd5c7bc
Signed-off-by: Joel King <joelking@codeaurora.org>