UPSTREAM: xhci: Fix bug after deq ptr set to link TRB.
commit 50d0206fcaea3e736f912fd5b00ec6233fb4ce44 upstream.

This patch fixes a particularly nasty bug that was revealed by the ring
expansion patches.  The bug has been present since the very beginning of
the xHCI driver history, and could have caused general protection faults
from bad memory accesses.

The first thing to note is that a Set TR Dequeue Pointer command can
move the dequeue pointer to a link TRB, if the canceled or stalled
transfer TD ended just before a link TRB.  The function to increment the
dequeue pointer, inc_deq, was written before cancellation and stall
support was added.  It assumed that the dequeue pointer could never
point to a link TRB.  It would unconditionally increment the dequeue
pointer at the start of the function, check if the pointer was now on a
link TRB, and move it to the top of the next segment if so.

This means that if a Set TR Dequeue Point command moved the dequeue
pointer to a link TRB, a subsequent call to inc_deq() would move the
pointer off the segment and into la-la-land.  It would then read from
that memory to determine if it was a link TRB.  Other functions would
often call inc_deq() until the dequeue pointer matched some other
pointer, which means this function would quite happily read all of
system memory before wrapping around to the right pointer value.

Often, there would be another endpoint segment from a different ring
allocated from the same DMA pool, which would be contiguous to the
segment inc_deq just stepped off of.  inc_deq would eventually find the
link TRB in that segment, and blindly move the dequeue pointer back to
the top of the correct ring segment.

The only reason the original code worked at all is because there was
only one ring segment.  With the ring expansion patches, the dequeue
pointer would eventually wrap into place, but the dequeue segment would
be out-of-sync.  On the second TD after the dequeue pointer was moved to
a link TRB, trb_in_td() would fail (because the dequeue pointer and
dequeue segment were out-of-sync), and this message would appear:

ERROR Transfer event TRB DMA ptr not part of current TD

This fixes bugzilla entry 4333 (option-based modem unhappy on USB 3.0
port: "Transfer event TRB DMA ptr not part of current TD", "rejecting
I/O to offline device"),

	https://bugzilla.kernel.org/show_bug.cgi?id=43333

and possibly other general protection fault bugs as well.

This patch should be backported to kernels as old as 2.6.31.  A separate
patch will be created for kernels older than 3.4, since inc_deq was
modified in 3.4 and this patch will not apply.

We were seeing below kernel crash:
[  158.665996] GPT: Use GNU Parted to correct GPT errors.
[  158.671143]  sda: sda1 sda2 sda3 sda4 sda5 sda6 sda7
[  158.683382] Unable to handle kernel paging request at virtual address f004e00c
[  158.689120] pgd = 80004000
[  158.691809] [f004e00c] *pgd=af006811, *pte=00000000, *ppte=00000000
[  158.698062] Internal error: Oops: 7 [#1] SMP ARM
[  158.702660] Modules linked in: nls_iso8859_1 nls_cp437 vfat fat i2c_dev uinput uvcvideo videobuf2_vmalloc rfcomm isl29018(C) industrialio(C) sbs_battery spidev btmrvl_sdi$
[  158.728961] CPU: 0    Tainted: G         C    (3.4.0 #6)
[  158.734260] PC is at inc_deq+0x110/0x130
[  158.738163] LR is at finish_td+0x154/0x238
[  158.742242] pc : [<80320ff8>]    lr : [<80323b90>]    psr: 80070193
[  158.742245] sp : 8072dd20  ip : ef39ca80  fp : 8072dd34
[  158.753698] r10: f0010510  r9 : 00000002  r8 : edaac300
[  158.758906] r7 : 00000000  r6 : eda52130  r5 : ef015000  r4 : ef39ca80
[  158.765416] r3 : f004e000  r2 : e83e9f00  r1 : edaac300  r0 : ef015000
[  158.771927] Flags: Nzcv  IRQs off  FIQs on  Mode SVC_32  ISA ARM  Segment kernel
[  158.779304] Control: 10c5387d  Table: addd006a  DAC: 00000015
[  158.785033] Process swapper/0 (pid: 0, stack limit = 0x8072c2f8)
[  158.791022] Stack: (0x8072dd20 to 0x8072e000)

BUG=chrome-os-partner:14992
TEST=Tested with 'Nobility NH-13' usb 3.0 device detected and
enumerated properly; could play an audio file present
in the device; No kernel crash after testing 20 times.

Change-Id: I3c689bd811fef6f91260e97cf6f1cb202ed899cc
Signed-off-by: Vikas C Sajjan <vikas.sajjan@samsung.com>
Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>
Tested-by: James Ettle <theholyettlz@googlemail.com>
Tested-by: Matthew Hall <mhall@mhcomputing.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 1aac2e73a8af6ce1cd6e938967dd17e2c8d38994)
Reviewed-on: https://gerrit.chromium.org/gerrit/34879
Tested-by: Alim Akhtar <alim.akhtar@samsung.com>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>