arm: exynos5: Fix the g3d clock hierarchy
As per Pg445 of Exynos_5250_SCP_UM_REV1.00, to setup an initial
frequency of 533MHz for the G3D block we need to have the parent
source clock of MOUTACLK_400_G3D as sclk_gpll and the divisor
register as 0. Fix the clock hierarchy for the same.

BUG=chrome-os-partner:11853
TEST=Build and boot Snow.
     1) Verify that ACLK400G3D print at bootup shows 533MHz.
     2) Verify CLK_SRC_TOP1[28], CLK_SRC_TOP2[28] and CLK_DIV_TOP0[24]
     bits are set appropriately for 533 MHz (1, 1, 0).
     3) Verify that "cat /sys/devices/platform/mali-t604.0/clock shows
     533 MHz (after applying Sam Leffler's Iae7ae43c: ARM MALI: Cleanup
     sysfs support.)

Change-Id: Ib1e9631bc1232f0df590ece5b9c5770aca1801e3
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/28489
Reviewed-by: Anush Elangovan <anush@google.com>
Commit-Ready: Anush Elangovan <anush@google.com>
Tested-by: Anush Elangovan <anush@google.com>
Reviewed-by: Doug Anderson <dianders@chromium.org>