ARM: EXYNOS5: dts: add power domain property to mfc/gscalar/sysmmu nodes
This patch adds the powerdomain property to the mfc/gsc/sysmmu nodes which
points to their respective power domain nodes. This property is retrieved
during the driver's probe and the device is added to its power domain.
MFC and its sysmmus are added to mfc_pd and all the gscalar devices and
their sysmmus are added to the gsc_pd.

BUG=chrome-os-partner:11793
Test=MFC/GSC/Sysmmu get suspended after probe and resume upon video
playback. Tested on Daisy.

Change-Id: Ifc811fb847e434b116b76dc85b908b6c0da187a5
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Signed-off-by: Naveen krishna Chatradhi <ch.naveen@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/29161
Tested-by: Prathyush Kalashwaram <prathyush@chromium.org>
Commit-Ready: Olof Johansson <olofj@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>