CHROMIUM: ALSA: hda/realtek - Add automic hook.
This hook mirrors the automute hook functionality, but for input
selection instead of output selection.  The hook will be used allow for
board specific mic routing behaviour.  On some machines the mic sense
signal is only valid if headphones are attached, this will allow for a
hook to filter mic sense events based on headphone jack state.

BUG=chrome-os-partner:11254
TEST=plug/unplug mic on parrot.

Change-Id: I01168baaa9ad9ee449eb8c2ac93e585f77f16065
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28844
Reviewed-by: Olof Johansson <olofj@chromium.org>