CHROMIUM: ALSA: hda/realtek - Ignore mic sense without headphones on Parrot.
Add an automic and automute hook for Parrot.  The mic hook checks if the
headphones are attached before checking the mic sense.  The mic sense
for Parrot reports "attached" when the combo jack is empty, this change
causes that to be ignored.  The automute hook is added so that the mic sense
will be checked when headphones state changes.

BUG=chrome-os-partner:11254
TEST=Call the echo bot with parrot plug/unplug headsets and mics.

Change-Id: Iee6c1a41c98b0829f6254ec2f9eacbead40d1468
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28869
Reviewed-by: Olof Johansson <olofj@chromium.org>