arm: cros5250-common: use always-on option for vdd regulator
The adc driver uses this regulator and it calls regulator_enable() and
regulator_disable().  Once it does that set of calls the vdd_1v8 rail
will turn off (although it was on at boot).  That's not so good since
the 1v8 rail is used by lots of components and needs to be on all the
time.

We'll change the regulator to be always on.

BUG=chrome-os-partner:11688
TEST=powerd_suspend then press power button; device wakes up

Change-Id: I7ef30b88746e86d331b06d927569f88f20d9593b
Signed-off-by: Naveen Krishna Chatradhi <ch.naveen@samsung.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28939
Reviewed-by: Olof Johansson <olofj@chromium.org>