msm: clock: Add clock mapping for rpm-regulator device
Add a CXO clock mapping for the rpm-regulator driver for all
targets which utilize a PMIC which requires CXO to be on in order
for its NLDOs to enable properly.  The workaround is only needed
if the target is utilizing a TCXO oscillator.

Change-Id: I6ce8356e0da71a713ddcd1f10e773b8c043a9bac
Signed-off-by: David Collins <collinsd@codeaurora.org>
(cherry picked from commit 6472d5bee58d09a02c88108e0f7169fcd00731fd)

Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>