mako: i2c: change max i2c clock to 384KHz
Devices support upto 400KHz but QUP I2C is recommended
to use 384KHz with the following message.

In the fast/standard mode, the I2C controller produces a
50/50 duty cycle clock. So at 400 kHz, the low clock
period time is 1.25 microseconds.
However, the I2C spec requires that the low time not be
shorter than 1.30 microseconds (the high need only be 0.60 microseconds).
To meet this requirement, the clock should be reduced to 384 kHz.

Change-Id: I2b2f567b6260ad2b2240b30cb2b8aec0b58a6af4