msm: acpuclock-krait: Don't change HFPLL rates unnecessarily
If two CPU speeds vote on the same L2 rate we unnecessarily set
the rate of the L2 PLL again even though the resulting rate of
the L2 will be the same. On 8960 for example, setting one CPU's
rate to 1512 KHz and another CPU's rate to 1458 KHz will cause
set_speed() to be called on the L2 scalable with a rate that is
the same as what the L2 clock is already programmed for.

Skip the rate switch in this case by bailing out early if the
target rate is the same as the current rate.

Change-Id: I6824a9ee1c74dcecd63d6b78983ce563eda1c4a9
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>