CHROMIUM: ALSA: hda - Don't resume unless needed.
When snd_hda_resume is called during system resume, don't resume the
codec unless it is needed.  If power_count is zero then the resume path
will be executed when it increments if the device is later opened.

BUG=chrome-os-partner:11780
TEST=power_Resume run with and without audio playing, also just after audio
stops playing.  Manually test on Lumpy with youtube audio and opening and
closing the lid.

Change-Id: I306e921c1ccbb1ce012a527bb9acbeba0b5ae5a8
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29537
Reviewed-by: Olof Johansson <olofj@chromium.org>