CHROMIUM: tpm: move resume to workqueue
This takes TPM off of critical path on system resume, allowing for a better
user experience.

Note that this work is only needed on some of the current versions of hardware
since the TPM loses power in suspend; that will be fixed later on so this
extra delay will no longer be needed at all.

BUG=chrome-os-partner:12065
TEST=power_Resume followed by kernel_TPMPing to make sure it came back as expected.

Change-Id: I338a214007b8f9df7414ca0f8246f50f9b3de746
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29576
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>