msm: 8974: Switch regulators from rpm to local control
Add devices for the PMIC that match against the qpnp-regulator
driver for all supported regulator devices from the apps side.

Remove most devices from the rpm_bus with the exception of the
devices that acpuclock is currently using. These need to stay
since acpuclock is calling rpm_regulator_get() that is only
available on the rpm-regulator driver.

For regulators that are duplicates between rpm and local, rename
the local names to end if a suffix of "_local" to make them
unique. With the absense of a RPM, these RPM devices will amount
to NOPs at runtime.

This switch to local control is temporary and will be reverted
once the RPM regulator support has been validated on 8974 CDP.

Change-Id: Ie50db44595e487dcad659e67d623d32f4883d987
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>