USB: dwc3: Adjust runtime pm the dwc3 driver to allow runtime suspend
The current code in the dwc3 probe effectively disables suspend from
ever working because it called a get() that was never put() until
device removal.  Change the runtime pm code to match the standard
formula and allow runtime pm to function.

Note that this doesn't enable full runtime pm on the DWC3 device in
that the port isn't put into a lower power mode when not used.
However it does allow users of dwc3 (like dwc3-exynos) to do some
amount of runtime power management.

BUG=chrome-os-partner:11066
TEST=With other patches in this series can run:
  echo auto > \
    /sys/devices/exynos-dwc3/dwc3.0/xhci-hcd/usb3/power/control
  echo auto > \
    /sys/devices/exynos-dwc3/dwc3.0/xhci-hcd/usb4/power/control
  grep phy_clock_en /sys/kernel/debug/gpio
...and see that the phy clock enable goes low when nothing is
plugged into the USB 3.0 port and goes high when something is
in the port.

Change-Id: I04ec77e68953afb4365605e6e40f49bd0ca9748e
Signed-off-by: Vivek Gautam <gautam.vivek@samsung.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30182
Reviewed-by: Olof Johansson <olofj@chromium.org>