CHROMIUM: chromeos_arm: Process ChromeOS GPIOs statically defined in the dt.
This change turns the chromeos_arm init code into a driver and device so that
it can have GPIOs attached to it by links which have nice, friendly names
instead of just numbers. The code which processes the GPIO is split out into
a separate function so that it can be reused in case other switches need the
same treatment.

BUG=chrome-os-partner:11297
TEST=Used update_kernel.sh to put onto a snow. Verified that the gpio showed
up as /sys/devices/platform/chromeos_arm/write-protect/.

Change-Id: Ie52a5b3dd751b05bd03c57aab316abe309821fbc
Signed-off-by: Gabe Black <gabeblack@google.com>
[olofj: cleanup, removed args, bindings update]
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30533