CHROMIUM: ARM: SAMSUNG: Add residue DMA operation.
Add a new dma op, residue.  It returns the count of bytes left to be
transfered by the dma.  This is useful for Audio, the Samsung dma ASoC
layer will use this to provide an accurate update to the hw_ptr that is
exported to user space.  ASoC wants large transfers to reduce the amount
of wakeups, but needs to be able to know how much audio has been played
for latency estimates.

BUG=chrome-os-partner:12278
TEST=aplay file, add debug prints to watch the residue value change.

Change-Id: I9783e0182f94ef2016d7f424219082005f8d07a0
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30614
Reviewed-by: Olof Johansson <olofj@chromium.org>
(cherry picked from commit f62c67258854112c65f25170578cca5775af0d4a)
Reviewed-on: https://gerrit.chromium.org/gerrit/32483