usb: dwc3: exynos5: Add support for powering off an external phy
On exynos we need an external PLL to pass compliance.  That PLL
can burn a lot of power, so provide a mechanism for turning it off
(and switching to a lower power clock) when in runtime suspend.

NOTE: This extends the phy interface that's already there for the
exynos dwc3 controller.  Because of some of the limitations of that
interface (mainly that there's no private data for the phy) we
needed to put some of the phy-related GPIO toggling in the main
dwc3-exynos glue.  That should be cleaned up in a future patch.

BUG=chrome-os-partner:11066
TEST=With other patches in this series can run:
  echo auto > \
    /sys/devices/exynos-dwc3/dwc3.0/xhci-hcd/usb3/power/control
  echo auto > \
    /sys/devices/exynos-dwc3/dwc3.0/xhci-hcd/usb4/power/control
  grep phy_clock_en /sys/kernel/debug/gpio
...and see that the phy clock enable goes low when nothing is
plugged into the USB 3.0 port and goes high when something is
in the port.

Change-Id: Ia5ccfbd640c8c07fbd3ead8bd6d962eb25532c4a
Signed-off-by: Vivek Gautam <gautam.vivek@samsung.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30734
Reviewed-by: Olof Johansson <olofj@chromium.org>