CHERRY-PICK: usb3: exynos5: Use external phy clock when needed; shut down when not
The usb3 phy clock is an external PLL that burns a lot of power but is
needed to pass USB 3.0 compliance testing.  When nothing is plugged
into the port we can turn the PLL off and switch to a different clock
to save power.  This other clock is enough to detect that a device
has been inserted.

We also do a little bit of cleanup to use register bit constants
introduced in a previous patch.

BUG=chrome-os-partner:11066
TEST=With other patches in this series can run:
  echo auto > \
    /sys/devices/exynos-dwc3/dwc3.0/xhci-hcd/usb3/power/control
  echo auto > \
    /sys/devices/exynos-dwc3/dwc3.0/xhci-hcd/usb4/power/control
  grep phy_clock_en /sys/kernel/debug/gpio
...and see that the phy clock enable goes low when nothing is
plugged into the USB 3.0 port and goes high when something is
in the port.

Change-Id: I0ca5ab7c8c06187251a8767768e682f43803bd7a
Signed-off-by: Vivek Gautam <gautam.vivek@samsung.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
(cherry picked from commit 92f37410f98833d8fb6610928c6a4ae3d7549d2b)
Reviewed-on: https://gerrit.chromium.org/gerrit/31059