drm/exynos: track vblank events on a per crtc basis
The goal of the change is to make sure we send the vblank event on the
current vblank. My hope is to fix any races that might be causing flicker.
After this change I only see a flicker in the transition plymouth and
X11.

Simplified the code by tracking vblank events on a per-crtc basis. This
allowed me to remove all error paths from the callback. It also allowed
me to remove the vblank wait from the callback.

BUG=chrome-os-partner:12586
TEST=Suspend/resume. Multipe VT Switches. Multiple LCD off/on. Some browsing.
TEST=Power down and restart multiple times.

Change-Id: Ia71863161c6cb7f2dca20d14a68acd5e19b0db53
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30781
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>