ASoC: msm: Add support timestamp update VCAP playback usecase
Compressed data in HDMI IN or VCAP usecase is propagated with
timestamp from DSP with each buffer. For playback, the same or
interpolated timestamp is updated with the corresponding buffer.
DSP will render the buffer a) without any delay if timestamp mode is
synchronous to absolute timestamp b) with a delay if timestamp is more
than the absolute timestamp c) drop the buffer if the timestamp is
less than the absolute timestamp. Add support for the same.

Change-Id: Iffa1143d468b2e8b1322aeb3f7b22c211b2916ad
Signed-off-by: Krishnankutty Kolathappilly <kkolat@codeaurora.org>