thermal: tsens8960: Fix delay in the ISR
There are more number of sensors on APQ8064 and MSM8930
whose status registers are to be read and checked
against enabled sensors if thresholds interrupts are
crossed. This causes delay in processing the ISR before
it is returned causing other timer events to be delayed.

Fix this by performing the status register reads for
the sensors in a seperate workqueue.

(cherry picked from commit 7dc2eb0488712eb9fc2c5ebfba56c0f453f6cdce)

Change-Id: I3c246110cd86889eb9336f29e18f98478373df96
CRs-Fixed: 390508
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>