msm: msm_bus: Add support for 64-bit bandwidth requests
With 8974, bandwidth requests are exceeding integer size.
Update ab and ib to ensure values over 4Gbps can be requested.

For 64-bit support, there are four changes in this patch:

Changes in the bus driver structures and functions to
update ab and ib values to uint64.

Updates in client drivers to ensure that the bus APIs
are called with the right types.

Updates to the device-tree. Device-tree nodes don't
provide a clean mechanism to read a combination of u32
and u64 variables within the same property name. So, the
ab and ib values must be specified in kbps instead of
bytes/sec from the device tree. The bus driver function
will convert the ab and ib values read from device-tree
to Mbps.

Change the existing property names to make them consistent
with the convention used by the new property  name, which
abides to the Device Tree convention.

Change-Id: Id37407c349a9cea17f6621d3bef2401b856f7b7d
CRs-Fixed: 408786
Signed-off-by: Gagan Mac <gmac@codeaurora.org>