CHERRY-PICK: ARM: EXYNOS5: add parent clock of gscalar
Currently there is no parent clock set for the gscalar devices
which results in the clock gating of gscalar to fail. This patch
sets the correct parent clock which ensures that gscalar gets
clock gated when not in use.

Procedure to check: (when gscl is not active)
Before: armio 0x10020920 -> fffff87f
After: armio 0x10020920 -> fffff870
where the last 4 bits show the clock gating status of the four
gscalar devices.

BUG=chrome-os-partner:13795
TEST=ran multimedia tests for audio, video and webgl.
And S2R (Open and close the lid) several times.

Change-Id: Iaaf410c14d0a2c5e262e5ef94085be8c55d41ab9
Original-Change-Id: Ic6e8b807469ab6fcc7fe8d2d56367958dde4ce62
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/31841
Tested-by: Prathyush Kalashwaram <prathyush@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Shaik Ameer Basha <shaik.ameer@samsung.com>
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Commit-Ready: Todd Broch <tbroch@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34183
Tested-by: naveen krishna chatradhi <naveen@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>