CHROMIUM: dts: Add BUCK8 operating mode bindings
BUCK8 can be turned off when the BUCK8EN bits of BUCK8CTRL register
are set to '01/10/00'. However, the register is set to '11' which
ignores the ENB8 signal. Set the operating mode for the BUCK to
obtain power savings by turning off the eMMC.

BUG=chrome-os-partner:12665
TEST=Verify BUCK8EN as "01". Mount, read and write the eMMC post resume.
Do a chromeos-install on the eMMC, boot from eMMC and check functioning
post resume. Suspend the system (rtcwake -m mem -s 25) and measure BUCK8
power. No significant change in power consumption.

Change-Id: I10768d46fd4fc6061e522fbcf570a5ddaa560114
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/31597
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Bryan Freed <bfreed@chromium.org>