gpio: msm: Dont re-enable the interrupt and clear the status
If a gpio interrupt is already enabled, do not re-enable it again
and do not clear the interrupt status.
Before this patch:
If an edge interrupt is marked as wakeup, the gpio sys suspend will
enable it.
On Resume - the gpio sys resume would clear the intr_status of the wakeup
irq.
Thus, if this particular wakeup irq triggered and woke us up, we lose its
intr_status before calling the gpio summary handler.
Now, when the summary handler is finally called, the irq handler for the
wakeup interrupt is never called.

With this change:
We are *Not* re-enabling the gpio-irq again to make sure that
when the irq framework unmasks all irqs on resume,
it does not clear the intr_status.

(cherry picked from commit f4f3658e26f0d1ba447ecd8e7bd17c5de40cd51c)

Change-Id: I66182f792271f3169760bb5f1caa03b35f8d2c72
CRs-Fixed: 400562
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>