CHROMIUM: Input: atmel_mxt_ts - disable input_dev when updating FW & Config
When the FW update is in progress, the linux input layer may
call open/close on the driver, which translates to writing
certain values into the device T9 control register, which does
not exist at all when the system is in bootloader mode and the
writing to register will make the firmware update corrupted.

So unregister the input device from input subsystem when doing
FW update and re-create it when update finishes. Also doing the
same for config update.

Signed-off-by: Yufeng Shen <miletus@chromium.org>

BUG=chromium-os:33981
TEST=Boot system and let the device auto-update the FW and make
     sure the update is successful.

Change-Id: If8f7ea2953f1a825cbcad4a548cf563345518c7d
Reviewed-on: https://gerrit.chromium.org/gerrit/31638
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Ready: Yufeng Shen <miletus@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>