drm/exynos: Clean-up IOMMU mapping creation order
The IOMMU mappings are shared across FIMD and HDMI. Due to the
ordering in the Makefile, the FIMD map call is executed first,
thus HDMI uses the parameters from there as well. The mapping size
passed in exynos_mixer.c is thus unused and misleading, so change
it to 0. Also, add sanity checks for that assumption in case
the ordering changes in the future.

BUG=None
TEST=None

Change-Id: Ib9c290ca4cdce72d7823fe3b4d584e0c2a7a9cd8
Signed-off-by: Pawel Osciak <posciak@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31898
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: Pawel Osciak <posciak@google.com>
Tested-by: Pawel Osciak <posciak@google.com>