qup_i2c: keep the ahb clock on for SGLTE 8960 only
Don't keep the AHB clock on for all types of 8960 targets
as the GSBI4 is shared by I2C and UART cores from different
processors on only SGLTE 8960 type platform. If the ahb clock
is kept on for GSBI4 from apps side, this needs a change in UART
driver on modem side to avoid touching the ahb clock and may
lead to some i2c errors if the modem doesn't have this change.
Hence we are keeping the ahb clock on for only SGLTE 8960 target
type.

CRs-fixed: 387908
Signed-off-by: Kiran Gunda <kgunda@codeaurora.org>
(cherry picked from commit af3b35627a9a3a09075bbeb0edc7876e20468367)

Change-Id: I49110fa7c4b7b5b72d7716603b9b3daeb1b17609
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>